added some comments and small tweaks

+++ matrix_kernel/src/memory/allocator.rs
@@ -13,7 +13,7 @@ pub(super) fn init_heap(
     frame_allocator: &mut impl FrameAllocator<Size4KiB>,
 ) -> Result<(), MapToError<Size4KiB>> {
     let page_range = {
-        let heap_start = VirtAddr::new(HEAP_START as u64);
+        let heap_start = VirtAddr::new(HEAP_START);
         let heap_end = heap_start + HEAP_SIZE - 1u64;
         let heap_start_page = Page::containing_address(heap_start);
         let heap_end_page = Page::containing_address(heap_end);
